Removed warning
git-svn-id: https://www.h3d.org:8090/H3DUtil/trunk@205 875a7764-6f23-0410-b93c-dbbdadb1ea73

+++ H3DUtil/src/DicomImage.cpp
@@ -198,7 +198,7 @@ void H3DUtil::DicomImage::loadImage( const vector< string > &urls ) {
                                     H3D_FULL_LOCATION );
     w = image->getWidth();
     h = image->getHeight();
-    d = urls.size();
+    d = (unsigned int) urls.size();
     bits_per_pixel = image->getDepth();
     if( !isPowerOfTwo( bits_per_pixel ) ) {
       bits_per_pixel = nextPowerOfTwo( bits_per_pixel );
